# [MM-29240] Improve buffer handling of UploadEmojiImage