@hanzei 2023-12-18
> Just a nit: can we refactor the benchmark to avoid the repetition of the for loop in every test? We could simply use a slice of testcases with the different `utils.CreateTestXYZ` and dimensions, and then iterate over those and apply the same loop body to them.

While that is feasible, I'm hesitant as it makes it harder to understand the side effects of the benchmarks. One important thing for benchmarks is that they actually test the performance of the code you want and are not tainted by side effects. (see also https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go)

In this case: Correctness > verbosity