@hanzei 2023-12-18
@lindy65 Would you please do some smoke testing around custom emoji uploads with different file sizes? No behaviour changes are expected with this PR.